uodated front end with seprated welcome page
