+++ src/index.js
@@ -1,14 +1,19 @@
 import React from 'react';
-import ReactDOM from 'react-dom/client'; // Note the change here
+import ReactDOM from 'react-dom/client';
+import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import App from './App';
-import './index.css'; // Optional: if you have global styles
+import UploadDoc from './UploadDoc';
+import './index.css';
 
-// Create a root
 const root = ReactDOM.createRoot(document.getElementById('root'));
 
-// Render the App component
 root.render(
   <React.StrictMode>
-    <App />
+    <Router>
+      <Routes>
+        <Route path="/" element={<App />} />
+        <Route path="/upload" element={<UploadDoc />} />
+      </Routes>
+    </Router>
   </React.StrictMode>
 );
